requirements.txt
annotated-types==0.7.0
coloredlogs==15.0.1
flatbuffers==24.3.25
humanfriendly==10.0
imageio==2.36.0
lazy_loader==0.4
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
onnxruntime==1.19.2
packaging==24.1
pillow==11.0.0
protobuf==5.28.3
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
python-dotenv==1.0.1
scikit-image==0.24.0
scipy==1.14.1
sympy==1.13.3
tifffile==2024.9.20
typing_extensions==4.12.2
